bookmarks changes:
The default(only!) drop behaviour in bookmarks is cut, this is not good. This means any file you drop on a directory in bookmarks will be moved to that location. I added the same behaviour as everything else in qtfm. Default is based on settings, if move is selected/default qtfm will still ask what to do if devices differ.