All: Port to Qt6
I ported only CMake. QMake is not modified. Code may (?) compile with
Qt5, but I am not dealing with it.

I run the code once. Code does seem to work, but I do know that
"opening" a file is no longer working. Copy does seem to work
(which means also move). Delete as well.

qtfm-tray does load, but does nothing. I did not look into it any more.

Code in Qt6 behaves better in Wayland and probably in HighDPI in
Windows/OSX - so this is another good reason to upgrade (and Qt5 is
EOL).